perf(npm): use direct binary for huge startup boost
