--- deployment/npm/install.js
@@ -1,4 +1,0 @@
-// @ts-check
-"use strict";
-
-require("./install_api").runInstall();
